Ignore Parcel's .cache; revamp README instructions
We really shouldn't be running a linking step (aka a "bundler", in our
case parcel) during development time; however this is out of the scope
of the refactor for now. We'll come back to this later when we remove
bs-webapi which, under the current setup, does require parcel.

We removed the yarn install instructions. Our community should ideally
standardize on a single package management CLI tool; unfortunately this
isn't a worthwhile fight for now, so we'll just softly recommend vanilla
npm over yarn if only to avoid bikeshedding (the author uses yarn; no
bias here).

A repo should ideally be as hermetic as possible; so we've added the
original tutorial's parcel commands in the README too
https://dev.to/jdeisenberg/manipulating-the-dom-with-rescript-3llf