Remove extra bs-webapi api usage
Use the same workflow as ad1fc9c and e5f7257.

Note the benefit of checking in out generated artifacts, especially in this diff: there is **no change in JS output**. Which means that our change is trivially provably correct (relative to the previous version).

As much as possible, you should leverage this property to quickly refactor code. I myself have refactored a total of more than 300K lines of code this way without causing a single new bug, and you can too.